[REFACTOR] Extract room background preview
